+++ packages/ra-materialui-treeview/src/getRecordFromNode.js
@@ -1,0 +1,1 @@
+export default ({ __children, __depth, ...record }) => record;
