+++ packages/ra-materialui-treeview/src/index.js
@@ -1,4 +1,4 @@
-export Treeview from './Treeview';
-export TreeviewNodeActions from './TreeviewNodeActions';
+export Tree from './Tree';
+export TreeNodeActions from './TreeNodeActions';
 export reducer from './reducers';
 export * from './actions';
